Ints -> Longs, Floats -> Doubles for better precision handle in real word scenario

+++ test-backend/src/main/java/test_backend/test_backend/Dto/SimulationDto.java
@@ -4,10 +4,10 @@ import java.util.Optional;
 
 public record SimulationDto(
         String N,
-        int P,
-        int I,
+        long P,
+        long I,
         int R,
-        float M,
+        double M,
         int Ti,
         int Tm,
         int Ts
