+++ test-backend/src/main/java/test_backend/test_backend/Entities/DailyResult.java
@@ -13,10 +13,10 @@ public class DailyResult {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private long ID;
-    private int Pi;
-    private int Pv;
-    private int Pm;
-    private int Pr;
+    private long Pi;
+    private long Pv;
+    private long Pm;
+    private long Pr;
 
     @ManyToOne
     private Simulation simulation;
